requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
faiss-cpu
sentence-transformers
google-generativeai
requests
